impl-node: remove old check
Remove the check for the target rate and duration. The drivers are
supposed to update that now.

The check is a little racy because a new target could have been
configured right after the driver emitted the ready signal. In that
case we should not update the clock values but let the driver run with
the old values and let the new values become active in the next cycle.