context: use timestamp for load-balancing of loops
We increment the refcount of the loop when we hand it out. We use this
refcount to do some load balancing. In some cases we also decrease the
refcount again when the release_loop() functions is called but we only do
this in some modules. The result is that in most cases the refcount
just keeps on going up and this is misleading.

Change the refcount to a last_used timestamp and prefer to use loops
that have not been used in a while.

Fixes #4436