adapter: pass follower direction to converter
Partially revert 86af9de7392cdb66948fc5ccced4e8619b57666c

The PortParam does not give enough information to derive the direction
of the converter. If the converter is configured in convert/convert
there is just no way to know when to output a quantum or not.

Fix this by doing a quick probe of the follower and then pass the
direction to the converter.

See !2227