context: handle transport updates better
The transport update is set in the node properties. If one client tries
to start and another tries to stop we have two conflicting desired states
in the nodes.

Fix this by making the state update a one time property and remove it
from the sever and client properties after updating it. We then need to
keep the new state around and apply it once.

Keep the node driver state as it is unless there was a transport state
update.

Fixes #4543